Add Feature: function getBooks() allows for retrieving all the books kept inside a database table. Added comments to better understand the structures managing the database.

+++ src/dbmanager.c
@@ -16,15 +16,33 @@
  *      - 2023-10-05: Worked on creating the database connect and close function.
  *                      Also defined the BookData struct.
  *      - 2023-10-07: Worked on creating addBook function.
- * 
+ *      - 2023-10-09: Finished the deleteBookById function and worked on getBooks function.
+ *      - 2023-10-11: Finished the getBooks function along with creating BookArray struct.
+ *                      Removed extra print statements, and added comments to functions.
 */
 
 #include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 #include "sqlite3.h"
 #include "dbmanager.h"
 
 static sqlite3* db;
+/**
+ * Creates the default tables
+*/
+static void createTable(void);
+/**
+ * For allocating memory to dest, the same size as src with null-terminator. Will
+ * print error to stderr if unable to allocate memory.
+ * 
+ * @param dest A pointer to a pointer to a character array. Memory will be allocated
+ *              to this pointer, the size of the character array src.
+ * @param src The source character array that contains the characters to be copied
+ * @returns 1 if operation was successful, else returns 0.
+*/
+static int copyField(char** dest, const char* src);
 
 int makeConnection(void) {
     if (db != NULL) {
@@ -40,7 +58,6 @@ int makeConnection(void) {
         return OPERATION_FAIL;
     }
 
-    printf("Database Connected!!\n");
     createTable();
     return OPERATION_SUCCESS;
 }
@@ -61,7 +78,7 @@ static void createTable(void) {
 
     int rc = sqlite3_exec(db, sqlStatement, 0, 0, &errorMsg);
     if (rc != SQLITE_OK) {
-        fprintf(stderr, "Creating Table Error: %s\n", errorMsg);
+        // Table is already created
         sqlite3_free(errorMsg);
         return;
     }
@@ -99,12 +116,15 @@ int addBook(BookData data) {
 
     // TODO store id in some sort of structure. First construct structure
     sqlite3_int64 lastRowID = sqlite3_last_insert_rowid(db);
-    printf("Data inserted successfully");
     sqlite3_finalize(stmt);
     return OPERATION_SUCCESS;
 }
 
-int removeBookById(int id) {
+int deleteBookById(int id) {
+    if (id < 1) {
+        return OPERATION_FAIL;
+    }
+
     int rc = 0;
     sqlite3_stmt* stmt;
     const char* sqlDelete = "DELETE FROM Books WHERE BookID = ?";
@@ -123,14 +143,98 @@ int removeBookById(int id) {
         return OPERATION_FAIL;
     }
 
-    printf("Data deleted successfully");
     sqlite3_finalize(stmt);
     return OPERATION_SUCCESS;
 }
 
-int removeBookByTitle(char* title) {
-    return 0;
+BookArray getBooks(void) {
+    int rc = 0;
+    sqlite3_stmt* stmt;
+    const char* sqlSelect = "SELECT * FROM Books";
+
+    // Return this error result if error
+    BookArray errorResult;
+    errorResult.books = NULL;
+    errorResult.count = -1;
+
+    BookData** books = NULL;
+    rc = sqlite3_prepare_v2(db, sqlSelect, -1, &stmt, 0);
+    if (rc != SQLITE_OK) {
+        fprintf(stderr, "SQL Error: %s\n", sqlite3_errmsg(db));
+        return errorResult;
+    }
+
+    int rowCount = 0;
+    while ((rc = sqlite3_step(stmt)) == SQLITE_ROW) {
+        books = realloc(books, (rowCount + 1) * sizeof(BookData *));
+        if (books == NULL) {
+            fprintf(stderr, "Error Allocating Memory in getBooks for books\n");
+            return errorResult;
+        }
+
+        books[rowCount] = malloc(sizeof(BookData));
+        if (books[rowCount] == NULL) {
+            fprintf(stderr, "Error Allocating Memory in getBooks for books of count\n");
+            freeBooks(books, rowCount);
+            return errorResult;
+        }
+
+        // Get book data from database and allocate memory for each field
+        if (!copyField(&(books[rowCount]->title), sqlite3_column_text(stmt, 1)) ||
+            !copyField(&(books[rowCount]->author), sqlite3_column_text(stmt, 2)) ||
+            !copyField(&(books[rowCount]->publisher), sqlite3_column_text(stmt, 3)) ||
+            !copyField(&(books[rowCount]->publicationDate), sqlite3_column_text(stmt, 4)) ||
+            !copyField(&(books[rowCount]->ISBN), sqlite3_column_text(stmt, 5)) ||
+            !copyField(&(books[rowCount]->genre), sqlite3_column_text(stmt, 6)) ||
+            !copyField(&(books[rowCount]->lang), sqlite3_column_text(stmt, 7))) {
+            fprintf(stderr, "Memory Allocation Error: %s\n", sqlite3_errmsg(db));
+            freeBooks(books, rowCount + 1); // rowCount plus one for including this row
+            return errorResult;
+        }
+
+        books[rowCount]->numPages = sqlite3_column_int(stmt, 8);
+        
+        rowCount++;
+    }
+
+    if (rc != SQLITE_DONE) {
+        fprintf(stderr, "Error In getBooks(): %s\n", sqlite3_errmsg(db));
+        freeBooks(books, rowCount);
+        return errorResult;
+    }
+
+    sqlite3_finalize(stmt);
+    BookArray result;
+    result.books = books;
+    result.count = rowCount;
+    return result;
 }
+
+static int copyField(char** dest, const char* src) {
+    *dest = malloc(strlen(src) + 1); // Plus one for null-terminator
+    if (*dest == NULL) {
+        fprintf(stderr, "Error Allocating Memory in getBooks\n");
+        return 0; // Return failure
+    }
+    // Copies database data into field
+    strcpy(*dest, src);
+    return 1; // Return success
+}
+
+void freeBooks(BookData** books, int numBooks) {
+    for (int i = 0; i < numBooks; i++) {
+        free(books[i]->title);
+        free(books[i]->author);
+        free(books[i]->publisher);
+        free(books[i]->publicationDate);
+        free(books[i]->ISBN);
+        free(books[i]->genre);
+        free(books[i]->lang);
+        free(books[i]);
+    }
+    free(books);
+}
+
 
 int closeConnection(void) {
     if (db == NULL) {
